+++ libs/common/src/lib/chart-helper.ts
@@ -1,16 +1,41 @@
 import { Chart, TooltipPosition } from 'chart.js';
+import { format } from 'date-fns';
 
-import { getBackgroundColor, getTextColor } from './helper';
-import { ColorScheme } from './types';
+import {
+  DATE_FORMAT,
+  DATE_FORMAT_MONTHLY,
+  DATE_FORMAT_YEARLY,
+  getBackgroundColor,
+  getTextColor
+} from './helper';
+import { ColorScheme, GroupBy } from './types';
+
+export function formatGroupedDate({
+  date,
+  groupBy
+}: {
+  date: Date;
+  groupBy: GroupBy;
+}) {
+  if (groupBy === 'month') {
+    return format(date, DATE_FORMAT_MONTHLY);
+  } else if (groupBy === 'year') {
+    return format(date, DATE_FORMAT_YEARLY);
+  }
+
+  return format(date, DATE_FORMAT);
+}
 
 export function getTooltipOptions({
   colorScheme,
   currency = '',
+  groupBy,
   locale = '',
   unit = ''
 }: {
   colorScheme?: ColorScheme;
   currency?: string;
+  groupBy?: GroupBy;
   locale?: string;
   unit?: string;
 } = {}) {
@@ -38,6 +63,13 @@ export function getTooltipOptions({
           }
         }
         return label;
+      },
+      title: (contexts) => {
+        if (groupBy) {
+          return formatGroupedDate({ groupBy, date: contexts[0].parsed.x });
+        }
+
+        return contexts[0].label;
       }
     },
     caretSize: 0,
@@ -96,4 +128,8 @@ export function getVerticalHoverLinePlugin(
     },
     id: 'verticalHoverLine'
   };
+}
+
+export function transformTickToAbbreviation(value: number) {
+  return value < 1000000 ? `${value / 1000}K` : `${value / 1000000}M`;
 }