+++ libs/ui/src/lib/fire-calculator/fire-calculator.component.ts
@@ -13,9 +13,11 @@ import {
   ViewChild
 } from '@angular/core';
 import { FormBuilder, FormControl } from '@angular/forms';
-import { getTooltipOptions } from '@ghostfolio/common/chart-helper';
+import {
+  getTooltipOptions,
+  transformTickToAbbreviation
+} from '@ghostfolio/common/chart-helper';
 import { primaryColorRgb } from '@ghostfolio/common/config';
-import { transformTickToAbbreviation } from '@ghostfolio/common/helper';
 import { ColorScheme } from '@ghostfolio/common/types';
 import {
   BarController,